requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiohttp-retry==2.8.3 ; python_version >= '3.7'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
coreapi==2.3.3
coreschema==0.0.4
django==4.2
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
openapi-codec==1.3.2
packaging==23.1 ; python_version >= '3.7'
pillow==9.5.0
psycopg2==2.9.6
pyjwt==2.6.0 ; python_version >= '3.7'
pytz==2023.3
requests==2.29.0 ; python_version >= '3.7'
ruamel.yaml==0.17.21 ; python_version >= '3'
ruamel.yaml.clib==0.2.7 ; python_version < '3.11' and platform_python_implementation == 'CPython'
simplejson==3.19.1 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
twilio==8.1.0
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.9.2 ; python_version >= '3.7'
